requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
redis==3.3.11
Werkzeug==0.16.0
